requirements.txt
customtkinter
tkintermapview
requests
pillow
pyinstaller